refactor: Move doc generation into its own package (test)
