+++ src/test/java/org/opentripplanner/generate/doc/BuildConfigUserGuideDocPrinter.java
@@ -1,17 +1,17 @@
-package org.opentripplanner.standalone.config.framework;
+package org.opentripplanner.generate.doc;
 
-import static org.opentripplanner.standalone.config.framework.JsonSupport.jsonNodeForTest;
-import static org.opentripplanner.standalone.config.framework.doc.DeprecatedParametersTable.createDeprecatedParametersTable;
-import static org.opentripplanner.standalone.config.framework.doc.ParameterDetailsList.listParametersWithDetails;
-import static org.opentripplanner.standalone.config.framework.doc.ParametersTable.createParametersTable;
+import static org.opentripplanner.generate.doc.framework.DeprecatedParametersTable.createDeprecatedParametersTable;
+import static org.opentripplanner.generate.doc.framework.ParameterDetailsList.listParametersWithDetails;
+import static org.opentripplanner.generate.doc.framework.ParametersTable.createParametersTable;
+import static org.opentripplanner.standalone.config.framework.JsonSupport.jsonNodeFromResource;
 
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.PrintStream;
 import java.util.Set;
+import org.opentripplanner.generate.doc.framework.MarkDownDocWriter;
 import org.opentripplanner.standalone.config.BuildConfig;
-import org.opentripplanner.standalone.config.framework.doc.MarkDownDocWriter;
 import org.opentripplanner.standalone.config.framework.json.NodeAdapter;
 import org.opentripplanner.standalone.config.framework.json.NodeInfo;
 
@@ -36,25 +36,9 @@ public class BuildConfigUserGuideDocPrinter {
       new MarkDownDocWriter(out)
     );
 
-    // We need to include all optional nested types we want to in the document, and
-    // at least one element need to added to nested objects as well (can be an unknown like {@code a:""})
-    // All required fields need to be included
-    var json =
-      """
-      {
-        boardingLocationTags : [ { a:1 } ],
-        dem : [ { source:1 } ],
-        elevationBucket : { accessKey:1, secretKey:1, bucketName:1 },
-        localFileNamePatterns : { a:1 },
-        netexDefaults : { a:1 },
-        osm : [{ source:1 }],
-        osmDefaults : { a:1 },
-        transferRequests : [{}],
-        transitFeeds : [{ type:1}]
-      }
-      """;
-
-    var conf = new BuildConfig(jsonNodeForTest(json), "build.config", false);
+    String source = "standalone/config/build-config.json";
+    var json = jsonNodeFromResource(source);
+    var conf = new BuildConfig(json, source, false);
     printer.write(conf.asNodeAdapter());
   }
