+++ src/test/java/org/opentripplanner/generate/doc/framework/ParameterDetailsList.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.standalone.config.framework.doc;
+package org.opentripplanner.generate.doc.framework;
 
 import java.util.function.Predicate;
 import org.opentripplanner.standalone.config.framework.json.NodeAdapter;
